removing imports, and change class names
