+++ cnxarchive/tests/views/test_in_book_search.py
@@ -5,12 +5,6 @@
 # Public License version 3 (AGPLv3).
 # See LICENCE.txt for details.
 # ###
-import os
-import datetime
-import glob
-import HTMLParser
-import time
-import json
 import unittest
 
 try:
@@ -18,16 +12,13 @@ try:
 except ImportError:
     import mock
 
-from pyramid import httpexceptions
 from pyramid import testing as pyramid_testing
-from pyramid.traversal import PATH_SAFE
 
 from ...utils import IdentHashShortId, IdentHashMissingVersion
 from .. import testing
 
 
-@mock.patch('cnxarchive.views.in_book_search.fromtimestamp', mock.Mock(side_effect=testing.mocked_fromtimestamp))
-class ViewsTestCase(unittest.TestCase):
+class InBookViewsTestCase(unittest.TestCase):
     fixture = testing.data_fixture
     maxDiff = 10000
