+++ cnxarchive/tests/views/test_recent.py
@@ -5,7 +5,6 @@
 # Public License version 3 (AGPLv3).
 # See LICENCE.txt for details.
 # ###
-import datetime
 import unittest
 
 try:
@@ -13,16 +12,12 @@ try:
 except ImportError:
     import mock
 
-from pyramid import httpexceptions
 from pyramid import testing as pyramid_testing
-from pyramid.traversal import PATH_SAFE
 
-from ...utils import IdentHashShortId, IdentHashMissingVersion
 from .. import testing
 
 
-@mock.patch('cnxarchive.views.recent.fromtimestamp', mock.Mock(side_effect=testing.mocked_fromtimestamp))
-class ViewsTestCase(unittest.TestCase):
+class RecentViewsTestCase(unittest.TestCase):
     fixture = testing.data_fixture
     maxDiff = 10000
