+++ cnxarchive/tests/views/test_search.py
@@ -6,9 +6,6 @@
 # See LICENCE.txt for details.
 # ###
 import os
-import datetime
-import glob
-import HTMLParser
 import time
 import json
 import unittest
@@ -18,12 +15,10 @@ try:
 except ImportError:
     import mock
 
-from pyramid import httpexceptions
 from pyramid import testing as pyramid_testing
 from pyramid.encode import url_quote
 from pyramid.traversal import PATH_SAFE
 
-from ...utils import IdentHashShortId, IdentHashMissingVersion
 from .. import testing
 
 
@@ -38,8 +33,7 @@ with open(SEARCH_RESULTS_FILEPATH, 'r') as file:
     SEARCH_RESULTS = json.load(file)
 
 
-@mock.patch('cnxarchive.views.search_view.fromtimestamp', mock.Mock(side_effect=testing.mocked_fromtimestamp))
-class ViewsTestCase(unittest.TestCase):
+class SearchViewsTestCase(unittest.TestCase):
     fixture = testing.data_fixture
     maxDiff = 10000
 
@@ -93,7 +87,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -111,7 +105,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -146,7 +140,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
 
         self.assertEqual(results['results']['total'], 0)
@@ -194,7 +188,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -215,7 +209,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -239,7 +233,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -263,7 +257,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -297,7 +291,7 @@ class ViewsTestCase(unittest.TestCase):
         # Test for no highlighting on specific field queries.
         self.request.params = {'q': 'title:"college physics"'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -331,7 +325,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -357,7 +351,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -383,7 +377,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -429,7 +423,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -475,7 +469,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -528,7 +522,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -547,7 +541,7 @@ class ViewsTestCase(unittest.TestCase):
         # Build the request
         self.request.params = {'q': '"college physics" type:module'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -569,7 +563,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -588,7 +582,7 @@ class ViewsTestCase(unittest.TestCase):
         # Build the request
         self.request.params = {'q': 'title:physics type:collection'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -607,7 +601,7 @@ class ViewsTestCase(unittest.TestCase):
         settings = self.settings.copy()
         settings['memcache-servers'] = ''
         config_kwargs = dict(settings=settings, request=self.request)
-        from ...views.search_view import search
+        from ...views.search import search
 
         # Build the request
         self.request.params = {'q': 'introduction',
@@ -672,7 +666,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -706,7 +700,7 @@ class ViewsTestCase(unittest.TestCase):
                                'per_page': '3',
                                'page': '2'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -735,7 +729,7 @@ class ViewsTestCase(unittest.TestCase):
                                'per_page': '3',
                                'page': '3'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -770,7 +764,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -783,7 +777,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.params = {'q': 'introduction',
                                'per_page': '3'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -797,7 +791,7 @@ class ViewsTestCase(unittest.TestCase):
                                'per_page': '3',
                                'nocache': 'True'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -814,7 +808,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -828,7 +822,7 @@ class ViewsTestCase(unittest.TestCase):
                                'per_page': '3',
                                'page': '2'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -845,7 +839,7 @@ class ViewsTestCase(unittest.TestCase):
                                'per_page': '3',
                                'page': '2'}
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -861,7 +855,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
@@ -894,7 +888,7 @@ class ViewsTestCase(unittest.TestCase):
         self.request.matched_route = mock.Mock()
         self.request.matched_route.name = 'search'
 
-        from ...views.search_view import search
+        from ...views.search import search
         results = search(self.request).json_body
         status = self.request.response.status
         content_type = self.request.response.content_type
