+++ cnxarchive/views/exceptions.py
@@ -6,31 +6,17 @@
 # See LICENCE.txt for details.
 # ###
 """Exception Views."""
-import os
-import json
 import logging
-from datetime import datetime, timedelta
 
-import psycopg2
-import psycopg2.extras
-from cnxepub.models import flatten_tree_to_ident_hashes
-from lxml import etree
-from pytz import timezone
 from pyramid import httpexceptions
-from pyramid.settings import asbool
-from pyramid.threadlocal import get_current_registry, get_current_request
 from pyramid.view import view_config
 
-from .. import config
-from ..database import SQL, get_tree, get_collated_content
 from ..utils import (
-    COLLECTION_MIMETYPE, IdentHashSyntaxError,
-    IdentHashShortId, IdentHashMissingVersion,
-    portaltype_to_mimetype, slugify, fromtimestamp,
-    join_ident_hash, split_ident_hash, split_legacy_hash
+    IdentHashSyntaxError, IdentHashShortId, IdentHashMissingVersion,
+    join_ident_hash, split_ident_hash
     )
-from .views_helpers import get_latest_version
-from .views_helpers import get_uuid
+from .helpers import get_latest_version
+from .helpers import get_uuid
 
 logger = logging.getLogger('cnxarchive')
