+++ cnxarchive/views/in_book_search.py
@@ -6,27 +6,20 @@
 # See LICENCE.txt for details.
 # ###
 """In Book Search Views."""
-import os
 import json
 import logging
-from datetime import datetime, timedelta
 
 import psycopg2
 import psycopg2.extras
-from cnxepub.models import flatten_tree_to_ident_hashes
-from lxml import etree
-from pytz import timezone
-from pyramid import httpexceptions
-from pyramid.settings import asbool
-from pyramid.threadlocal import get_current_registry, get_current_request
+from pyramid.threadlocal import get_current_registry
 from pyramid.view import view_config
 
 from .. import config
-from ..database import SQL, get_tree, get_collated_content
+from ..database import SQL
 from ..utils import (
-    IdentHashShortId, IdentHashMissingVersion, split_ident_hash, fromtimestamp
+    IdentHashShortId, IdentHashMissingVersion, split_ident_hash
     )
-from .views_helpers import get_uuid
+from .helpers import get_uuid
 
 logger = logging.getLogger('cnxarchive')
