+++ cnxarchive/views/search.py
@@ -6,38 +6,21 @@
 # See LICENCE.txt for details.
 # ###
 """Search Views."""
-import os
 import json
 import logging
-from datetime import datetime, timedelta
-from re import compile
 
 import psycopg2
 import psycopg2.extras
-from cnxepub.models import flatten_tree_to_ident_hashes
-from lxml import etree
-from pytz import timezone
-from pyramid import httpexceptions
-from pyramid.settings import asbool
-from pyramid.threadlocal import get_current_registry, get_current_request
+from pyramid.threadlocal import get_current_registry
 from pyramid.view import view_config
 
 from .. import config
 from .. import cache
-# FIXME double import
-from .. import database
-from ..database import SQL, get_tree, get_collated_content
+from ..database import SQL
 from ..search import (
     DEFAULT_PER_PAGE, QUERY_TYPES, DEFAULT_QUERY_TYPE,
     Query,
     )
-from ..utils import (
-    COLLECTION_MIMETYPE, IdentHashSyntaxError,
-    IdentHashShortId, IdentHashMissingVersion,
-    portaltype_to_mimetype, slugify, fromtimestamp,
-    join_ident_hash, split_ident_hash, split_legacy_hash
-    )
-
 
 logger = logging.getLogger('cnxarchive')
 _BLOCK_SIZE = 4096 * 64  # 256K
